requirements.txt
sentence-transformers
faiss-cpu
# Add other dependencies if needed later
